Enum nodle_chain::cli::Subcommand

pub enum Subcommand {
    Key(KeySubcommand),
    Benchmark(BenchmarkCmd),
    Verify(VerifyCmd),
    Vanity(VanityCmd),
    Sign(SignCmd),
    BuildSpec(BuildSpecCmd),
    CheckBlock(CheckBlockCmd),
    ExportBlocks(ExportBlocksCmd),
    ExportState(ExportStateCmd),
    ImportBlocks(ImportBlocksCmd),
    PurgeChain(PurgeChainCmd),
    Revert(RevertCmd),
}

Possible subcommands of the main binary.

Variants

Key(KeySubcommand)

Key management cli utilities

Tuple Fields of Key

0: KeySubcommand

Benchmark(BenchmarkCmd)

The custom benchmark subcommmand benchmarking runtime pallets.

Tuple Fields of Benchmark

0: BenchmarkCmd

Verify(VerifyCmd)

Verify a signature for a message, provided on STDIN, with a given (public or secret) key.

Tuple Fields of Verify

0: VerifyCmd

Vanity(VanityCmd)

Generate a seed that provides a vanity address.

Tuple Fields of Vanity

0: VanityCmd

Sign(SignCmd)

Sign a message, with a given (secret) key.

Tuple Fields of Sign

0: SignCmd

BuildSpec(BuildSpecCmd)

Build a chain specification.

Tuple Fields of BuildSpec

0: BuildSpecCmd

CheckBlock(CheckBlockCmd)

Validate blocks.

Tuple Fields of CheckBlock

0: CheckBlockCmd

ExportBlocks(ExportBlocksCmd)

Export blocks.

Tuple Fields of ExportBlocks

0: ExportBlocksCmd

ExportState(ExportStateCmd)

Export the state of a given block into a chain spec.

Tuple Fields of ExportState

0: ExportStateCmd

ImportBlocks(ImportBlocksCmd)

Import blocks.

Tuple Fields of ImportBlocks

0: ImportBlocksCmd

PurgeChain(PurgeChainCmd)

Remove the whole chain.

Tuple Fields of PurgeChain

0: PurgeChainCmd

Revert(RevertCmd)

Revert the chain to a previous state.

Tuple Fields of Revert

0: RevertCmd