Module nodle_chain::para_service

Re-exports

pub use eden_executor::*;

Modules

eden_executor

Structs

ParaId

Unique identifier of a parachain.

Functions

eden_parachain_build_import_queue

Build the import queue for the rococo parachain runtime.

new_partial

Starts a ServiceBuilder for a full service.

start_eden_parachain_node

Start a Eden parachain node.

start_node_impl

Start a node with the given parachain Configuration and relay chain Configuration.

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Balance

Balance of an account.

Block

Block type.

Hash

A hash of some data used by the chain.

Header

Header type.

Nonce

Index of a transaction in the chain.